# Conflicts:
#	src/Client/Command/Commands/SkinStealCommand.cpp

+++ src/Client/Module/Modules/Misc/ScriptMarketplace/ScriptMarketplace.hpp
@@ -1,12 +1,33 @@
 #pragma once
 
+#include <random>
 #include <winrt/impl/windows.storage.2.h>
 #include <wininet.h>
 #include "../../../../Client.hpp"
 
 class ScriptMarketplace : public Listener {
-private:
-    std::string GetString(const std::string &URL) {
+
+public:
+    ScriptMarketplace() {
+        Listen(this, ProtocolEvent, &ScriptMarketplace::onProtocol);
+        Listen(this, ProtocolEvent, &ScriptMarketplace::onProtocolConfig);
+    }
+
+    ~ScriptMarketplace() {
+        Deafen(this, ProtocolEvent, &ScriptMarketplace::onProtocol);
+        Deafen(this, ProtocolEvent, &ScriptMarketplace::onProtocolConfig);
+    }
+
+    std::string scriptPath() {
+        using namespace winrt::Windows::Storage;
+
+        auto roamingFolder = Utils::getRoamingPath();
+        std::filesystem::path scriptsDir = std::filesystem::path(roamingFolder) / "Flarial" / "scripts";
+
+        return scriptsDir.string();
+    }
+
+    static std::string GetString(const std::string &URL) {
         try {
             HINTERNET interwebs = InternetOpenA("Samsung Smart Fridge", INTERNET_OPEN_TYPE_DIRECT, NULL, NULL, 0);
             if (!interwebs) {
@@ -31,23 +52,6 @@ private:
         }
         return "";
     }
-public:
-    ScriptMarketplace() {
-        Listen(this, ProtocolEvent, &ScriptMarketplace::onProtocol);
-    }
-
-    ~ScriptMarketplace() {
-        Deafen(this, ProtocolEvent, &ScriptMarketplace::onProtocol);
-    }
-
-    std::string scriptPath() {
-        using namespace winrt::Windows::Storage;
-
-        auto roamingFolder = Utils::getRoamingPath();
-        std::filesystem::path scriptsDir = std::filesystem::path(roamingFolder) / "Flarial" / "scripts";
-
-        return scriptsDir.string();
-    }
 
     void onProtocol(ProtocolEvent event) {
         if (event.getPath() == std::wstring(L"flarial-scripting")) {
@@ -57,16 +61,54 @@ public:
                     Logger::info("script id {}", id);
                     std::string url = "http://node2.sear.host:5019/api/scripts/" + id + "/download";
                     std::string data = GetString(url);
-                    std::ofstream file(Utils::getRoamingPath() + "\\Flarial\\tmpd.tmp", std::ios::binary);
+                    std::ofstream file(Utils::getClientPath() + "\\Scripts\\Modules\\tmpd.tmp", std::ios::binary);
                     Logger::info("data: {}", data.c_str());
                     file.write(data.c_str(), data.size());
                     file.close();
-                    Utils::extractFromFile(Utils::getRoamingPath() + "\\Flarial\\tmpd.tmp", scriptPath() + "\\" + id);
-                    std::filesystem::remove(Utils::getRoamingPath() + "\\Flarial\\tmpd.tmp");
 
                     ModuleManager::restartModules = true;
                 }
             }
         }
+    }
+
+    void onProtocolConfig(ProtocolEvent event) {
+
+        if (event.getPath() == std::wstring(L"flarial-configs")) {
+            for (const auto &pair: event.getProtocolArgs()) {
+                if (pair.first == std::wstring(L"scriptId")) {
+                    std::string id = String::WStrToStr(pair.second);
+                    Logger::info("script id {}", id);
+                    std::string url = "http://node2.sear.host:5019/api/configs/" + id + "/download";
+                    std::string data = ScriptMarketplace::GetString(url);
+                    std::ofstream file(Utils::getRoamingPath() + "\\Flarial\\tmpd.tmp", std::ios::binary);
+                    Logger::info("data: {}", data.c_str());
+                    file.write(data.c_str(), data.size());
+                    file.close();
+                    std::random_device rd;
+                    std::mt19937 gen(rd());
+
+                    std::uniform_int_distribution<> distrib(1000, 9000);
+
+                    int random_number = distrib(gen);
+                    std::string configname = "config-" + std::to_string(random_number);
+                    Client::createConfig(configname);
+                    std::string to = Utils::getRoamingPath() + "\\Flarial\\Config\\" + configname;
+
+                    Utils::extractFromFile(Utils::getRoamingPath() + "\\Flarial\\tmpd.tmp", to);
+                    std::filesystem::remove(Utils::getRoamingPath() + "\\Flarial\\tmpd.tmp");
+
+                    FlarialGUI::Notify("Imported Config as: " + configname);
+
+                    reloadAllConfigs();
+                }
+            }
+        }
+    }
+
+    static void reloadAllConfigs() {
+        Client::availableConfigs.clear();
+        Client::loadAvailableConfigs();
+        ModuleManager::restartModules = true;
     }
 };
